# Optimize BeEquivalentTo