--- a/.gitignore
+++ b/.gitignore
@@ -185,3 +185,6 @@ docs/node_modules
 
 # Documentation file needed for StyleCop.Analyzers
 Tests/FluentAssertions.Specs/FluentAssertions.Specs.xml
+
+# BenchmarkDotNet
+Tests/Benchmarks/BenchmarkDotNet.Artifacts/