@jnyrup 2021-08-22
Unsure what you're asking for. Are you looking for a summary of the data from the benchmarks I included?