@jnyrup 2021-08-22
In terms of cpu the two commits are rather equal whereas for memory the type check of Subject gives the most benefits.